requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
pydantic
mlflow
python-dotenv==1.0.1
pytest
httpx